Task 3 upd
